rrhh/requirements.txt
appdirs==1.4.3
Django==1.7.5
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.6.1
django-filter==0.9.2
django-sortable==0.3.1
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pyzk==0.5
reportlab==3.4.0
six==1.10.0
Unipath==1.1
